Fixes and updates

+++ src/WidgetTree/nuiTopLevel.cpp
@@ -1139,7 +1139,7 @@ NGL_TOUCHES_DEBUG( NGL_OUT(_T("CallMouseClick [%d] BEGIN\n"), rInfo.TouchId) );
     }
     
     bool hook = false;
-    for (size_t i = containers.size() - 1; i >= 0 && !hook; i--)
+    for (int i = containers.size() - 1; i >= 0 && !hook; i--)
     {
       nglMouseInfo info(rInfo);
       containers[i]->GlobalToLocal(info.X, info.Y);
@@ -1189,7 +1189,7 @@ void nuiTopLevel::UpdateMouseCursor(const nuiWidgetList& rWidgets)
     }
   }
   
-  size_t i = 0;
+  int i = 0;
   while (i < rWidgets.size())
   {
     nuiWidget* pWidget = rWidgets[i];
@@ -1209,7 +1209,7 @@ void nuiTopLevel::DispatchKeyboardFocus(const nuiWidgetList& rWidgets)
 {
   CheckValid();
   nuiAutoRef;
-  size_t i = 0;
+  int i = 0;
   
   while (i < rWidgets.size())
   {
@@ -1289,7 +1289,7 @@ NGL_TOUCHES_DEBUG( NGL_OUT(_T("CallMouseUnclick [%d] BEGIN\n"), rInfo.TouchId) )
     }
     
     bool hook = false;
-    for (size_t i = containers.size() - 1; i >= 0 && !hook; i--)
+    for (int i = (int)containers.size() - 1; i >= 0 && !hook; i--)
     {
       nglMouseInfo info(rInfo);
       containers[i]->GlobalToLocal(info.X, info.Y);
@@ -1472,7 +1472,8 @@ NGL_TOUCHES_DEBUG( NGL_OUT(_T("nuiTopLevel::CallMouseMove X:%d Y:%d\n"), rInfo.X
     }
     
     bool hook = false;
-    for (size_t i = containers.size() - 1; i >= 0 && !hook; i--)
+      
+    for (int i = (int)containers.size() - 1; i >= 0 && !hook; i--)
     {
       nglMouseInfo info(rInfo);
       containers[i]->GlobalToLocal(info.X, info.Y);
@@ -1626,7 +1627,7 @@ bool nuiTopLevel::CallMouseWheel (nglMouseInfo& rInfo)
     }
 
     bool hook = false;
-    for (size_t i = containers.size() - 1; i >= 0 && !hook; i--)
+    for (int i = containers.size() - 1; i >= 0 && !hook; i--)
     {
       nglMouseInfo info(rInfo);
       containers[i]->GlobalToLocal(info.X, info.Y);
@@ -1984,7 +1985,7 @@ bool nuiTopLevel::DrawTree(class nuiDrawContext *pContext)
     // Prepare the layout changes:
     pContext->ResetState();
 
-    size_t count = mDirtyRects.size();
+    int count = mDirtyRects.size();
 
 //    NGL_OUT("drawing %d partial rects\n", count);
